fix: corrige configuração do NextAuth movendo para arquivo separado
